Fix type from Integer to Long

+++ src/main/java/com/kylerriggs/kanban/priority/PriorityRepository.java
@@ -2,5 +2,5 @@ package com.kylerriggs.kanban.priority;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PriorityRepository extends JpaRepository<Priority, Integer> {
+public interface PriorityRepository extends JpaRepository<Priority, Long> {
 }
